Organized satDNA and miRNA + moved data to NAS
