--- multi_mappings/analysis/FBgn0025525/FBgn0025525_roi.bed
@@ -1,1 +1,0 @@
-chr3L	1140429	1177276	-
