--- multi_mappings/analysis/FBgn0025525/FBgn0025525_roibeds.txt
@@ -1,82 +1,0 @@
-bin252.bed
-bin253.bed
-bin254.bed
-bin255.bed
-bin256.bed
-bin257.bed
-bin258.bed
-bin263.bed
-bin264.bed
-bin265.bed
-bin267.bed
-bin271.bed
-bin272.bed
-bin273.bed
-bin274.bed
-bin275.bed
-bin278.bed
-bin279.bed
-bin281.bed
-bin282.bed
-bin283.bed
-bin284.bed
-bin285.bed
-bin286.bed
-bin287.bed
-bin288.bed
-bin289.bed
-bin290.bed
-bin291.bed
-bin292.bed
-bin293.bed
-bin294.bed
-bin295.bed
-bin296.bed
-bin297.bed
-bin298.bed
-bin299.bed
-bin300.bed
-bin301.bed
-bin302.bed
-bin303.bed
-bin304.bed
-bin305.bed
-bin306.bed
-bin307.bed
-bin308.bed
-bin309.bed
-bin310.bed
-bin311.bed
-bin312.bed
-bin313.bed
-bin314.bed
-bin315.bed
-bin316.bed
-bin317.bed
-bin318.bed
-bin319.bed
-bin320.bed
-bin321.bed
-bin322.bed
-bin323.bed
-bin324.bed
-bin325.bed
-bin326.bed
-bin327.bed
-bin328.bed
-bin329.bed
-bin330.bed
-bin331.bed
-bin332.bed
-bin333.bed
-bin334.bed
-bin335.bed
-bin336.bed
-bin337.bed
-bin338.bed
-bin339.bed
-bin340.bed
-bin341.bed
-bin342.bed
-bin343.bed
-bin344.bed
