--- multi_mappings/analysis/H4/FBgn0040765_histone_reads.bed
@@ -1,27 +1,0 @@
-chr2R	11001546	11001570	SRR1187947.24716002	48	-
-chr2R	11001546	11001572	SRR1187947.16929849	48	+
-chr2R	11001547	11001571	SRR1187947.9659672	48	+
-chr2R	11001547	11001572	SRR1187947.2506502	50	-
-chr2R	11001547	11001575	SRR1187947.1491574	50	+
-chr2R	11001548	11001571	SRR1187947.7437209	46	-
-chr2R	11001548	11001572	SRR1187947.7068140	48	-
-chr2R	11001549	11001573	SRR1187947.1491574	48	+
-chr2R	11001549	11001575	SRR1187947.9659672	52	+
-chr2R	11001550	11001573	SRR1187947.7437209	46	-
-chr2R	11072990	11073015	SRR1187947.9659672	50	-
-chr2R	11072991	11073014	SRR1187947.7437209	46	+
-chr2R	11072991	11073015	SRR1187947.1491574	48	-
-chr2R	11072991	11073015	SRR1187947.2506502	48	+
-chr2R	11072992	11073023	SRR1187947.2506502	48	+
-chr2R	11072993	11073016	SRR1187947.7437209	46	+
-chr2R	11072994	11073018	SRR1187947.7068140	48	+
-chr2R	11072994	11073019	SRR1187947.2506502	50	+
-chr2R	11072995	11073018	SRR1187947.7437209	46	+
-chr2R	11072995	11073019	SRR1187947.1491574	48	-
-chr2R	11072995	11073019	SRR1187947.9659672	48	-
-chr2R	11072997	11073020	SRR1187947.7437209	46	+
-chr2R	11072998	11073022	SRR1187947.7068140	48	+
-chr2R	11072998	11073023	SRR1187947.2506502	50	+
-chr2R	11072999	11073022	SRR1187947.7437209	46	+
-chr2R	11072999	11073023	SRR1187947.1491574	48	-
-chr2R	11072999	11073023	SRR1187947.9659672	48	-
