--- multi_mappings/analysis/H4/FBgn0040765_histone_reads.txt
@@ -1,27 +1,0 @@
-SRR1187947.24716002
-SRR1187947.16929849
-SRR1187947.9659672
-SRR1187947.2506502
-SRR1187947.1491574
-SRR1187947.7437209
-SRR1187947.7068140
-SRR1187947.1491574
-SRR1187947.9659672
-SRR1187947.7437209
-SRR1187947.9659672
-SRR1187947.7437209
-SRR1187947.1491574
-SRR1187947.2506502
-SRR1187947.2506502
-SRR1187947.7437209
-SRR1187947.7068140
-SRR1187947.2506502
-SRR1187947.7437209
-SRR1187947.1491574
-SRR1187947.9659672
-SRR1187947.7437209
-SRR1187947.7068140
-SRR1187947.2506502
-SRR1187947.7437209
-SRR1187947.1491574
-SRR1187947.9659672
