--- multi_mappings/analysis/H4/FBgn0040765_roi.bed
@@ -1,2 +1,0 @@
-chr2R	10977956	11116378
-
