--- multi_mappings/analysis/H4/fbn_0040765_histone.piProfile.txt
@@ -1,30 +1,0 @@
-1	0	0.000000	0.000000
-2	0	0.000000	0.000000
-3	0	0.000000	0.000000
-4	0	0.000000	0.000000
-5	0	0.000000	0.000000
-6	0	0.000000	0.000000
-7	0	0.000000	0.000000
-8	0	0.000000	0.000000
-9	0	0.000000	0.000000
-10	0	0.000000	0.000000
-11	0	0.000000	0.000000
-12	0	0.000000	0.000000
-13	0	0.000000	0.000000
-14	0	0.000000	0.000000
-15	0	0.000000	0.000000
-16	0	0.000000	0.000000
-17	0	0.000000	0.000000
-18	0	0.000000	0.000000
-19	0	0.000000	0.000000
-20	0	0.000000	0.000000
-21	0	0.000000	0.000000
-22	0	0.000000	0.000000
-23	0	0.000000	0.000000
-24	0	0.000000	0.000000
-25	0	0.000000	0.000000
-26	0	0.000000	0.000000
-27	0	0.000000	0.000000
-28	0	0.000000	0.000000
-29	0	0.000000	0.000000
-30	0	0.000000	0.000000
