--- multi_mappings/analysis/H4/h4_histone_reads.bed
@@ -1,16 +1,0 @@
-chr2L	21529678	21529706	SRR1187947.48670	56	-
-chr2L	21529682	21529711	SRR1187947.24455983	58	-
-chr2L	21529705	21529732	SRR1187947.17851986	54	-
-chr2L	21529714	21529739	SRR1187947.8403018	50	-
-chr2L	21529717	21529742	SRR1187947.7068140	50	+
-chr2L	21529717	21529744	SRR1187947.16929849	50	-
-chr2L	21529718	21529743	SRR1187947.2506502	50	+
-chr2L	21529719	21529742	SRR1187947.7437209	46	+
-chr2L	21529719	21529743	SRR1187947.1491574	48	-
-chr2L	21529719	21529743	SRR1187947.9659672	48	-
-chr2L	21529720	21529744	SRR1187947.24716002	48	+
-chr2L	21529730	21529754	SRR1187947.13656532	48	-
-chr2L	21529756	21529769	SRR1187947.23757221	54	-
-chr2L	21529762	21529769	SRR1187947.3526002	53	-
-chr2L	21529767	21529769	SRR1187947.3562121	50	-
-chr2L	21529767	21529769	SRR1187947.17573543	56	-
