--- multi_mappings/analysis/H4/h4_histone_reads.txt
@@ -1,16 +1,0 @@
-SRR1187947.48670
-SRR1187947.24455983
-SRR1187947.17851986
-SRR1187947.8403018
-SRR1187947.7068140
-SRR1187947.16929849
-SRR1187947.2506502
-SRR1187947.7437209
-SRR1187947.1491574
-SRR1187947.9659672
-SRR1187947.24716002
-SRR1187947.13656532
-SRR1187947.23757221
-SRR1187947.3526002
-SRR1187947.3562121
-SRR1187947.17573543
