--- multi_mappings/analysis/H4/h4_roi.bed
@@ -1,1 +1,0 @@
-chr2L	21529242	21529769
