--- multi_mappings/analysis/fbgn_0027339/fbgn_0027339_roi.bed
@@ -1,2 +1,0 @@
-chr3L	22789233	22835194
-
