--- multi_mappings/annotation_data/extremly_highovary.txt
@@ -1,48 +1,0 @@
-#SUBMITTED ID	LOCATION_ARM	LOCATION_MAX	LOCATION_MIN	LOCATION_STRAND	NAME
-FBgn0005585	3R	9630851	9628691	-1	Calreticulin
-FBgn0020618	2L	7827315	7825347	-1	Receptor of activated protein kinase C 1
-FBgn0003941	2L	4228561	4227699	1	Ribosomal protein L40
-FBgn0034138	2R	16571230	16570112	-1	Ribosomal protein S15
-FBgn0003983	2L	13411631	13411141	-1	Vitelline membrane 34Ca
-FBgn0000559	2L	21682164	21677913	-1	eukaryotic translation elongation factor 2
-FBgn0011726	2R	24046805	24044453	-1	twinstar
-FBgn0000405	2R	22806948	22803521	-1	Cyclin B
-FBgn0002593	2L	420864	419952	1	Ribosomal protein LP1
-FBgn0011761	X	5313125	5312171	1	deadhead
-FBgn0039359	3R	25685051	25683623	-1	Ribosomal protein L27
-FBgn0019936	3R	20823356	20822181	1	Ribosomal protein S20
-FBgn0003979	2L	5960340	5959596	-1	Vitelline membrane 26Aa
-FBgn0020238	3R	18250550	18242531	1	14-3-3epsilon
-FBgn0001226	3L	9385694	9384063	1	Heat shock protein 27
-FBgn0000150	3R	31746077	31745172	-1	abnormal wing discs
-FBgn0004362	2R	21716727	21701458	1	High mobility group protein D
-FBgn0264922	2L	6967644	6966780	-1	smt3
-FBgn0003279	3R	27924888	27922821	1	Ribosomal protein L4
-FBgn0000360	X	8483311	8481837	1	Chorion protein 38
-FBgn0013269	3R	15333133	15331584	-1	FK506-binding protein 39kD
-FBgn0020255	X	11148716	11144852	-1	Ran
-FBgn0026372	3L	1668654	1666288	-1	Ribosomal protein L23A
-FBgn0004419	2L	10242172	10239341	1	maternal expression at 31B
-FBgn0285952	2R	24059483	24057798	1	eukaryotic translation elongation factor 5
-FBgn0266599	3R	15246614	15242687	1	Heat shock protein cognate 4
-FBgn0284245	2R	11895762	11891135	1	eukaryotic translation elongation factor 1 alpha 1
-FBgn0032987	2L	22247875	22247111	-1	Ribosomal protein L21
-FBgn0002579	X	629483	628338	-1	Ribosomal protein L36
-FBgn0003942	2L	10410456	10408352	1	Ribosomal protein S27A
-FBgn0001233	3L	3197059	3192969	1	Heat shock protein 83
-FBgn0000042	X	5905399	5900861	1	Actin 5C
-FBgn0034802	2R	22937077	22934569	-1	CCHC-type zinc finger nucleic acid binding protein
-FBgn0066084	2R	24903935	24903416	-1	Ribosomal protein L41
-FBgn0002622	3R	23360207	23358812	1	Ribosomal protein S3
-FBgn0004045	X	10055498	10053811	1	Yolk protein 1
-FBgn0284243	2R	19451600	19447319	-1	beta-Tubulin at 56D
-FBgn0003980	2L	5957639	5957004	1	Vitelline membrane 26Ab
-FBgn0003884	3R	7088839	7086599	1	alpha-Tubulin at 84B
-FBgn0045823	3L	9717716	9714324	1	visgun
-FBgn0086472	3R	11216287	11215028	-1	Ribosomal protein S25
-FBgn0014026	X	6533348	6530897	-1	Ribosomal protein L7A
-FBgn0017579	3L	8602392	8601006	1	Ribosomal protein L14
-FBgn0032400	2L	12046118	12045343	-1	-
-FBgn0039757	3R	30209020	30206447	-1	Ribosomal protein S7
-FBgn0039857	3R	31596204	31594792	-1	Ribosomal protein L6
-FBgn0041775	3L	12520954	12515059	1	trailer hitch
